Pull index routing into strategy object (backport of #77211)
This pulls the calculation of the shard id for an (id, routing) pair
into a little strategy class, `IndexRouting`. This is easier to test and
should be easier to extend.

My hope is that this is an incremental readability improvement. My
ulterior motive is that this is where I want to land our new
routing-by-dimensions work for tsdb.
