+++ server/src/test/java/org/elasticsearch/cluster/routing/OperationRoutingTests.java
@@ -7,10 +7,8 @@
  */
 package org.elasticsearch.cluster.routing;
 
-import org.elasticsearch.Version;
 import org.elasticsearch.action.support.replication.ClusterStateCreationUtils;
 import org.elasticsearch.cluster.ClusterState;
-import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.cluster.routing.allocation.decider.AwarenessAllocationDecider;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.settings.ClusterSettings;
@@ -26,13 +24,11 @@ import org.elasticsearch.threadpool.TestThreadPool;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.TreeMap;
 
 import static org.elasticsearch.cluster.routing.OperationRouting.IGNORE_AWARENESS_ATTRIBUTES_DEPRECATION_MESSAGE;
 import static org.hamcrest.CoreMatchers.containsString;
@@ -42,314 +38,6 @@ import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.object.HasToString.hasToString;
 
 public class OperationRoutingTests extends ESTestCase {
-    public void testGenerateShardId() {
-        int[][] possibleValues = new int[][] {
-            {8,4,2}, {20, 10, 2}, {36, 12, 3}, {15,5,1}
-        };
-        for (int i = 0; i < 10; i++) {
-            int[] shardSplits = randomFrom(possibleValues);
-            assertEquals(shardSplits[0], (shardSplits[0] / shardSplits[1]) * shardSplits[1]);
-            assertEquals(shardSplits[1], (shardSplits[1] / shardSplits[2]) * shardSplits[2]);
-            IndexMetadata metadata = IndexMetadata.builder("test").settings(settings(Version.CURRENT)).numberOfShards(shardSplits[0])
-                .numberOfReplicas(1).build();
-            String term = randomAlphaOfLength(10);
-            final int shard = OperationRouting.generateShardId(metadata, term, null);
-            IndexMetadata shrunk = IndexMetadata.builder("test").settings(settings(Version.CURRENT)).numberOfShards(shardSplits[1])
-                .numberOfReplicas(1)
-                .setRoutingNumShards(shardSplits[0]).build();
-            int shrunkShard = OperationRouting.generateShardId(shrunk, term, null);
-
-            Set<ShardId> shardIds = IndexMetadata.selectShrinkShards(shrunkShard, metadata, shrunk.getNumberOfShards());
-            assertEquals(1, shardIds.stream().filter((sid) -> sid.id() == shard).count());
-
-            shrunk = IndexMetadata.builder("test").settings(settings(Version.CURRENT)).numberOfShards(shardSplits[2]).numberOfReplicas(1)
-                .setRoutingNumShards(shardSplits[0]).build();
-            shrunkShard = OperationRouting.generateShardId(shrunk, term, null);
-            shardIds = IndexMetadata.selectShrinkShards(shrunkShard, metadata, shrunk.getNumberOfShards());
-            assertEquals(Arrays.toString(shardSplits), 1, shardIds.stream().filter((sid) -> sid.id() == shard).count());
-        }
-    }
-
-    public void testGenerateShardIdSplit() {
-        int[][] possibleValues = new int[][] {
-            {2,4,8}, {2, 10, 20}, {3, 12, 36}, {1,5,15}
-        };
-        for (int i = 0; i < 10; i++) {
-            int[] shardSplits = randomFrom(possibleValues);
-            assertEquals(shardSplits[0], (shardSplits[0] * shardSplits[1]) / shardSplits[1]);
-            assertEquals(shardSplits[1], (shardSplits[1] * shardSplits[2]) / shardSplits[2]);
-            IndexMetadata metadata = IndexMetadata.builder("test").settings(settings(Version.CURRENT)).numberOfShards(shardSplits[0])
-                .numberOfReplicas(1).setRoutingNumShards(shardSplits[2]).build();
-            String term = randomAlphaOfLength(10);
-            final int shard = OperationRouting.generateShardId(metadata, term, null);
-            IndexMetadata split = IndexMetadata.builder("test").settings(settings(Version.CURRENT)).numberOfShards(shardSplits[1])
-                .numberOfReplicas(1)
-                .setRoutingNumShards(shardSplits[2]).build();
-            int shrunkShard = OperationRouting.generateShardId(split, term, null);
-
-            ShardId shardId = IndexMetadata.selectSplitShard(shrunkShard, metadata, split.getNumberOfShards());
-            assertNotNull(shardId);
-            assertEquals(shard, shardId.getId());
-
-            split = IndexMetadata.builder("test").settings(settings(Version.CURRENT)).numberOfShards(shardSplits[2]).numberOfReplicas(1)
-                .setRoutingNumShards(shardSplits[2]).build();
-            shrunkShard = OperationRouting.generateShardId(split, term, null);
-            shardId = IndexMetadata.selectSplitShard(shrunkShard, metadata, split.getNumberOfShards());
-            assertNotNull(shardId);
-            assertEquals(shard, shardId.getId());
-        }
-    }
-
-    public void testPartitionedIndex() {
-        // make sure the same routing value always has each _id fall within the configured partition size
-        for (int shards = 1; shards < 5; shards++) {
-            for (int partitionSize = 1; partitionSize == 1 || partitionSize < shards; partitionSize++) {
-                IndexMetadata metadata = IndexMetadata.builder("test")
-                    .settings(settings(Version.CURRENT))
-                    .numberOfShards(shards)
-                    .routingPartitionSize(partitionSize)
-                    .numberOfReplicas(1)
-                    .build();
-
-                for (int i = 0; i < 20; i++) {
-                    String routing = randomUnicodeOfLengthBetween(1, 50);
-                    Set<Integer> shardSet = new HashSet<>();
-
-                    for (int k = 0; k < 150; k++) {
-                        String id = randomUnicodeOfLengthBetween(1, 50);
-
-                        shardSet.add(OperationRouting.generateShardId(metadata, id, routing));
-                    }
-
-                    assertEquals(partitionSize, shardSet.size());
-                }
-            }
-        }
-    }
-
-    public void testPartitionedIndexShrunk() {
-        Map<String, Map<String, Integer>> routingIdToShard = new HashMap<>();
-
-        Map<String, Integer> routingA = new HashMap<>();
-        routingA.put("a_0", 1);
-        routingA.put("a_1", 2);
-        routingA.put("a_2", 2);
-        routingA.put("a_3", 2);
-        routingA.put("a_4", 1);
-        routingA.put("a_5", 2);
-        routingIdToShard.put("a", routingA);
-
-        Map<String, Integer> routingB = new HashMap<>();
-        routingB.put("b_0", 0);
-        routingB.put("b_1", 0);
-        routingB.put("b_2", 0);
-        routingB.put("b_3", 0);
-        routingB.put("b_4", 3);
-        routingB.put("b_5", 3);
-        routingIdToShard.put("b", routingB);
-
-        Map<String, Integer> routingC = new HashMap<>();
-        routingC.put("c_0", 1);
-        routingC.put("c_1", 1);
-        routingC.put("c_2", 0);
-        routingC.put("c_3", 0);
-        routingC.put("c_4", 0);
-        routingC.put("c_5", 1);
-        routingIdToShard.put("c", routingC);
-
-        Map<String, Integer> routingD = new HashMap<>();
-        routingD.put("d_0", 2);
-        routingD.put("d_1", 2);
-        routingD.put("d_2", 3);
-        routingD.put("d_3", 3);
-        routingD.put("d_4", 3);
-        routingD.put("d_5", 3);
-        routingIdToShard.put("d", routingD);
-
-        IndexMetadata metadata = IndexMetadata.builder("test")
-                .settings(settings(Version.CURRENT))
-                .setRoutingNumShards(8)
-                .numberOfShards(4)
-                .routingPartitionSize(3)
-                .numberOfReplicas(1)
-                .build();
-
-        for (Map.Entry<String, Map<String, Integer>> routingIdEntry : routingIdToShard.entrySet()) {
-            String routing = routingIdEntry.getKey();
-
-            for (Map.Entry<String, Integer> idEntry : routingIdEntry.getValue().entrySet()) {
-                String id = idEntry.getKey();
-                int shard = idEntry.getValue();
-
-                assertEquals(shard, OperationRouting.generateShardId(metadata, id, routing));
-            }
-        }
-    }
-
-    public void testPartitionedIndexBWC() {
-        Map<String, Map<String, Integer>> routingIdToShard = new HashMap<>();
-
-        Map<String, Integer> routingA = new HashMap<>();
-        routingA.put("a_0", 3);
-        routingA.put("a_1", 2);
-        routingA.put("a_2", 2);
-        routingA.put("a_3", 3);
-        routingIdToShard.put("a", routingA);
-
-        Map<String, Integer> routingB = new HashMap<>();
-        routingB.put("b_0", 5);
-        routingB.put("b_1", 0);
-        routingB.put("b_2", 0);
-        routingB.put("b_3", 0);
-        routingIdToShard.put("b", routingB);
-
-        Map<String, Integer> routingC = new HashMap<>();
-        routingC.put("c_0", 4);
-        routingC.put("c_1", 4);
-        routingC.put("c_2", 3);
-        routingC.put("c_3", 4);
-        routingIdToShard.put("c", routingC);
-
-        Map<String, Integer> routingD = new HashMap<>();
-        routingD.put("d_0", 3);
-        routingD.put("d_1", 4);
-        routingD.put("d_2", 4);
-        routingD.put("d_3", 4);
-        routingIdToShard.put("d", routingD);
-
-        IndexMetadata metadata = IndexMetadata.builder("test")
-                .settings(settings(Version.CURRENT))
-                .numberOfShards(6)
-                .routingPartitionSize(2)
-                .numberOfReplicas(1)
-                .build();
-
-        for (Map.Entry<String, Map<String, Integer>> routingIdEntry : routingIdToShard.entrySet()) {
-            String routing = routingIdEntry.getKey();
-
-            for (Map.Entry<String, Integer> idEntry : routingIdEntry.getValue().entrySet()) {
-                String id = idEntry.getKey();
-                int shard = idEntry.getValue();
-
-                assertEquals(shard, OperationRouting.generateShardId(metadata, id, routing));
-            }
-        }
-    }
-
-    /**
-     * Ensures that all changes to the hash-function / shard selection are BWC
-     */
-    public void testBWC() {
-        Map<String, Integer> termToShard = new TreeMap<>();
-        termToShard.put("sEERfFzPSI", 1);
-        termToShard.put("cNRiIrjzYd", 7);
-        termToShard.put("BgfLBXUyWT", 5);
-        termToShard.put("cnepjZhQnb", 3);
-        termToShard.put("OKCmuYkeCK", 6);
-        termToShard.put("OutXGRQUja", 5);
-        termToShard.put("yCdyocKWou", 1);
-        termToShard.put("KXuNWWNgVj", 2);
-        termToShard.put("DGJOYrpESx", 4);
-        termToShard.put("upLDybdTGs", 5);
-        termToShard.put("yhZhzCPQby", 1);
-        termToShard.put("EyCVeiCouA", 1);
-        termToShard.put("tFyVdQauWR", 6);
-        termToShard.put("nyeRYDnDQr", 6);
-        termToShard.put("hswhrppvDH", 0);
-        termToShard.put("BSiWvDOsNE", 5);
-        termToShard.put("YHicpFBSaY", 1);
-        termToShard.put("EquPtdKaBZ", 4);
-        termToShard.put("rSjLZHCDfT", 5);
-        termToShard.put("qoZALVcite", 7);
-        termToShard.put("yDCCPVBiCm", 7);
-        termToShard.put("ngizYtQgGK", 5);
-        termToShard.put("FYQRIBcNqz", 0);
-        termToShard.put("EBzEDAPODe", 2);
-        termToShard.put("YePigbXgKb", 1);
-        termToShard.put("PeGJjomyik", 3);
-        termToShard.put("cyQIvDmyYD", 7);
-        termToShard.put("yIEfZrYfRk", 5);
-        termToShard.put("kblouyFUbu", 7);
-        termToShard.put("xvIGbRiGJF", 3);
-        termToShard.put("KWimwsREPf", 4);
-        termToShard.put("wsNavvIcdk", 7);
-        termToShard.put("xkWaPcCmpT", 0);
-        termToShard.put("FKKTOnJMDy", 7);
-        termToShard.put("RuLzobYixn", 2);
-        termToShard.put("mFohLeFRvF", 4);
-        termToShard.put("aAMXnamRJg", 7);
-        termToShard.put("zKBMYJDmBI", 0);
-        termToShard.put("ElSVuJQQuw", 7);
-        termToShard.put("pezPtTQAAm", 7);
-        termToShard.put("zBjjNEjAex", 2);
-        termToShard.put("PGgHcLNPYX", 7);
-        termToShard.put("hOkpeQqTDF", 3);
-        termToShard.put("chZXraUPBH", 7);
-        termToShard.put("FAIcSmmNXq", 5);
-        termToShard.put("EZmDicyayC", 0);
-        termToShard.put("GRIueBeIyL", 7);
-        termToShard.put("qCChjGZYLp", 3);
-        termToShard.put("IsSZQwwnUT", 3);
-        termToShard.put("MGlxLFyyCK", 3);
-        termToShard.put("YmscwrKSpB", 0);
-        termToShard.put("czSljcjMop", 5);
-        termToShard.put("XhfGWwNlng", 1);
-        termToShard.put("cWpKJjlzgj", 7);
-        termToShard.put("eDzIfMKbvk", 1);
-        termToShard.put("WFFWYBfnTb", 0);
-        termToShard.put("oDdHJxGxja", 7);
-        termToShard.put("PDOQQqgIKE", 1);
-        termToShard.put("bGEIEBLATe", 6);
-        termToShard.put("xpRkJPWVpu", 2);
-        termToShard.put("kTwZnPEeIi", 2);
-        termToShard.put("DifcuqSsKk", 1);
-        termToShard.put("CEmLmljpXe", 5);
-        termToShard.put("cuNKtLtyJQ", 7);
-        termToShard.put("yNjiAnxAmt", 5);
-        termToShard.put("bVDJDCeaFm", 2);
-        termToShard.put("vdnUhGLFtl", 0);
-        termToShard.put("LnqSYezXbr", 5);
-        termToShard.put("EzHgydDCSR", 3);
-        termToShard.put("ZSKjhJlcpn", 1);
-        termToShard.put("WRjUoZwtUz", 3);
-        termToShard.put("RiBbcCdIgk", 4);
-        termToShard.put("yizTqyjuDn", 4);
-        termToShard.put("QnFjcpcZUT", 4);
-        termToShard.put("agYhXYUUpl", 7);
-        termToShard.put("UOjiTugjNC", 7);
-        termToShard.put("nICGuWTdfV", 0);
-        termToShard.put("NrnSmcnUVF", 2);
-        termToShard.put("ZSzFcbpDqP", 3);
-        termToShard.put("YOhahLSzzE", 5);
-        termToShard.put("iWswCilUaT", 1);
-        termToShard.put("zXAamKsRwj", 2);
-        termToShard.put("aqGsrUPHFq", 5);
-        termToShard.put("eDItImYWTS", 1);
-        termToShard.put("JAYDZMRcpW", 4);
-        termToShard.put("lmvAaEPflK", 7);
-        termToShard.put("IKuOwPjKCx", 5);
-        termToShard.put("schsINzlYB", 1);
-        termToShard.put("OqbFNxrKrF", 2);
-        termToShard.put("QrklDfvEJU", 6);
-        termToShard.put("VLxKRKdLbx", 4);
-        termToShard.put("imoydNTZhV", 1);
-        termToShard.put("uFZyTyOMRO", 4);
-        termToShard.put("nVAZVMPNNx", 3);
-        termToShard.put("rPIdESYaAO", 5);
-        termToShard.put("nbZWPWJsIM", 0);
-        termToShard.put("wRZXPSoEgd", 3);
-        termToShard.put("nGzpgwsSBc", 4);
-        termToShard.put("AITyyoyLLs", 4);
-        IndexMetadata metadata = IndexMetadata.builder("test").settings(settings(Version.CURRENT)).numberOfShards(8)
-            .numberOfReplicas(1).build();
-        for (Map.Entry<String, Integer> entry : termToShard.entrySet()) {
-            String key = entry.getKey();
-            int shard = randomBoolean() ?
-                OperationRouting.generateShardId(metadata, key, null) : OperationRouting.generateShardId(metadata, "foobar", key);
-            assertEquals(shard, entry.getValue().intValue());
-        }
-    }
-
     public void testPreferNodes() throws InterruptedException, IOException {
         TestThreadPool threadPool = null;
         ClusterService clusterService = null;
